requirements.txt
numpy<2
torch==2.1.0
coremltools==7.1
open_clip_torch
pillow
tqdm
onnx==1.14.0
onnx-coreml
